refactor: 엔티티 연관관계 수정  (bugpigg/issue-tracker#1)
