+++ BE/src/main/java/com/team09/issue_tracker/comment/Emogi.java
@@ -1,6 +1,7 @@
 package com.team09.issue_tracker.comment;
 
 import com.team09.issue_tracker.comment.Comment;
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -13,6 +14,7 @@ public class Emogi {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "emogi_id")
 	private Long id;
 
 	private String unicodeValue;
