+++ BE/src/main/java/com/team09/issue_tracker/label/Label.java
@@ -1,5 +1,6 @@
 package com.team09.issue_tracker.label;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -10,6 +11,7 @@ public class Label {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "label_id")
 	private Long id;
 
 	private String title;
