+++ BE/src/main/java/com/team09/issue_tracker/member/Member.java
@@ -1,5 +1,6 @@
 package com.team09.issue_tracker.member;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
@@ -12,6 +13,7 @@ public class Member {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "member_id")
 	private Long id;
 
 	@Enumerated(EnumType.STRING)
