+++ BE/src/main/java/com/team09/issue_tracker/milestone/Milestone.java
@@ -1,6 +1,7 @@
 package com.team09.issue_tracker.milestone;
 
 import java.time.LocalDate;
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -11,6 +12,7 @@ public class Milestone {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "milestone_id")
 	private Long id;
 
 	private String title;
